Refactored single byte hex formatter
